(refactor) O3-2685: Inconsistent / Duplicated Appointments List UI between Home App and Appointments App (#1084)
