--- packages/esm-appointments-app/src/appointments/forms/cancel-form/cancel-appointment.scss
@@ -1,19 +1,0 @@
-@use '@carbon/styles/scss/spacing';
-
-.formContainer {
-  margin: spacing.$spacing-05;
-}
-
-.inputContainer {
-  margin: spacing.$spacing-05;
-}
-
-.buttonContainer {
-  display: flex;
-  & > button {
-    width: 50%;
-    height: spacing.$spacing-10;
-    display: flex;
-    align-items: flex-start;
-  }
-}
