+++ packages/esm-appointments-app/src/appointments/scheduled/appointments-list.component.tsx
@@ -2,14 +2,15 @@ import React from 'react';
 import { useAppointmentList } from '../../hooks/useAppointmentList';
 import AppointmentsTable from '../common-components/appointments-table.component';
 import { filterByServiceType } from '../utils';
+import dayjs from 'dayjs';
 
-interface AppointmentsByStatusProps {
+interface AppointmentsListProps {
   appointmentServiceType?: string;
-  status: string;
+  status?: string;
   title: string;
   date: string;
 }
-const AppointmentsByStatus: React.FC<AppointmentsByStatusProps> = ({ appointmentServiceType, status, title, date }) => {
+const AppointmentsList: React.FC<AppointmentsListProps> = ({ appointmentServiceType, status, title, date }) => {
   const { appointmentList, isLoading } = useAppointmentList(status, date);
 
   const appointments = filterByServiceType(appointmentList, appointmentServiceType).map((appointment, index) => {
@@ -22,4 +23,4 @@ const AppointmentsByStatus: React.FC<AppointmentsByStatusProps> = ({ appointment
   return <AppointmentsTable appointments={appointments} isLoading={isLoading} tableHeading={title} />;
 };
 
-export default AppointmentsByStatus;
+export default AppointmentsList;
