--- packages/esm-appointments-app/src/home/appointments-base-table.scss
@@ -1,360 +1,0 @@
-@use '@carbon/colors';
-@use '@carbon/layout';
-@use '@carbon/type';
-
-.container {
-  background-color: colors.$white-0;
-  border: 1px solid colors.$gray-20;
-  width: 100%;
-  margin: 0 auto;
-  max-width: 95vw;
-  margin-top: layout.$spacing-09;
-  margin-bottom: layout.$spacing-12;
-}
-
-.loadingContainer {
-  @extend .container;
-  border: none;
-}
-
-.headerContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: layout.$spacing-04 0 layout.$spacing-04 layout.$spacing-05;
-  background-color: colors.$white-0;
-}
-
-.backgroundDataFetchingIndicator {
-  align-items: center;
-  display: flex;
-  flex: 1 1 0%;
-  justify-content: center;
-}
-
-.tableContainer section {
-  position: relative;
-}
-
-.hiddenRow {
-  display: none;
-}
-
-.tableContainer {
-  a {
-    text-decoration: none;
-  }
-
-  th {
-    color: colors.$gray-70;
-  }
-
-  :global(.cds--data-table) {
-    background-color: colors.$gray-20;
-  }
-
-  .tableToolbar {
-    min-height: 0;
-    top: -22px;
-  }
-
-  .tableToolbarContent {
-    height: layout.$spacing-07;
-    margin-bottom: layout.$spacing-02;
-  }
-}
-
-.tableToolbar {
-  min-height: 0;
-  top: -22px;
-}
-
-.startTime {
-  font-weight: 700;
-  margin-right: layout.$spacing-03;
-}
-
-.nameContainer {
-  display: flex;
-}
-
-.serviceContainer {
-  display: flex;
-  align-items: center;
-  width: max-content;
-}
-
-.serviceColor {
-  border-radius: 50%;
-  margin-right: layout.$spacing-03;
-  width: layout.$spacing-05;
-  height: layout.$spacing-05;
-  background-color: colors.$green-50;
-}
-
-.pagination {
-  overflow: hidden;
-
-  &:global(.cds--pagination) {
-    border-top: none;
-  }
-}
-
-.hiddenRow {
-  display: none;
-}
-
-.emptyRow {
-  padding: 0 layout.$spacing-05;
-  display: flex;
-  align-items: center;
-}
-
-.expandedAppointmentsRow > td > div {
-  max-height: max-content;
-}
-
-.expandedAppointmentsRow td {
-  padding: 0 layout.$spacing-07;
-}
-
-.expandedAppointmentsRow th[colspan] td[colspan] > div:first-child {
-  padding: 0 layout.$spacing-05;
-}
-
-.completeIcon,
-.missedIcon,
-.cancelIcon {
-  width: 8rem;
-  display: flex;
-  align-items: center;
-  height: layout.$spacing-09;
-
-  svg {
-    margin-left: layout.$spacing-03;
-  }
-}
-
-.completeIcon {
-  color: colors.$green-50;
-  svg {
-    fill: colors.$green-50;
-  }
-}
-
-.missedIcon {
-  color: colors.$red-50;
-  svg {
-    fill: colors.$red-50;
-  }
-}
-
-.tabletHeading {
-  h4 {
-    @include type.type-style('heading-03');
-    color: colors.$gray-70;
-  }
-}
-
-.desktopHeading,
-.tabletHeading {
-  display: flex;
-  justify-content: space-between;
-  text-align: left;
-  text-transform: capitalize;
-
-  h4 {
-    @include type.type-style('heading-compact-02');
-    color: colors.$gray-70;
-
-    &:after {
-      content: '';
-      display: block;
-      width: 2rem;
-      padding-top: 3px;
-      border-bottom: 0.375rem solid var(--brand-03);
-    }
-  }
-}
-
-.tileContainer {
-  background-color: colors.$white-0;
-  border-top: 1px solid colors.$gray-20;
-  padding: layout.$spacing-11 0;
-}
-
-.tile {
-  text-align: center;
-}
-
-.tileContent {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.content {
-  @include type.type-style('heading-compact-01');
-  color: colors.$gray-70;
-  margin-top: layout.$spacing-05;
-  margin-bottom: layout.$spacing-03;
-}
-
-.helper {
-  @include type.type-style('body-compact-01');
-  color: colors.$gray-70;
-}
-
-.separator {
-  @include type.type-style('body-compact-02');
-  color: colors.$gray-70;
-  width: 80%;
-  margin: layout.$spacing-06 auto;
-  overflow: hidden;
-  text-align: center;
-
-  &::before,
-  &::after {
-    background-color: colors.$gray-40;
-    content: '';
-    display: inline-block;
-    height: 1px;
-    position: relative;
-    vertical-align: middle;
-    width: 50%;
-  }
-
-  &::before {
-    right: layout.$spacing-03;
-    margin-left: -50%;
-  }
-
-  &::after {
-    left: layout.$spacing-03;
-    margin-right: -50%;
-  }
-}
-
-.actionLink {
-  margin: -1rem;
-}
-
-.actionLinks {
-  display: flex;
-  align-items: center;
-  margin-top: -0.75rem;
-}
-
-.menuItemLink {
-  text-decoration: none;
-}
-
-.menuItem {
-  max-width: none;
-}
-
-.seeAllLink {
-  position: relative;
-  left: 22rem;
-  top: -8.5rem;
-  min-height: layout.$spacing-07;
-  padding: layout.$spacing-03;
-}
-
-.filterLabelContainer {
-  display: flex;
-}
-
-.filterLabel {
-  background-color: colors.$white-0;
-  padding: 0.02rem layout.$spacing-05;
-  border-radius: layout.$spacing-05;
-  margin-right: layout.$spacing-03;
-  :hover {
-    background-color: colors.$white-0;
-  }
-}
-
-.viewCalendarButton {
-  margin: -1.25rem 0;
-}
-
-.divider {
-  width: 1px;
-  height: layout.$spacing-05;
-  color: colors.$gray-20;
-  margin: 0 layout.$spacing-05;
-}
-
-.overflowMenu {
-  vertical-align: middle !important;
-}
-
-// Overriding styles for RTL support
-html[dir='rtl'] {
-  .headerContainer {
-    padding: layout.$spacing-04 layout.$spacing-05 layout.$spacing-04 0;
-    svg {
-      margin-left: 0;
-      margin-right: layout.$spacing-03;
-    }
-    h4 {
-      text-align: right;
-    }
-  }
-
-  .pagination {
-    & > :nth-child(1) {
-      & > span {
-        margin-left: 0;
-        margin-right: layout.$spacing-05;
-      }
-      & > :nth-child(2) {
-        select {
-          border-right: none;
-          border-left: 1px solid var(--cds-border-subtle);
-        }
-      }
-    }
-    & > :nth-child(2) {
-      & > :first-child {
-        select {
-          border-left: none;
-        }
-      }
-      & > span {
-        margin-right: 0.0625rem;
-        margin-left: layout.$spacing-05;
-      }
-      & > :last-child {
-        & > :first-child {
-          border-right: 1px solid var(--cds-border-subtle);
-        }
-        svg {
-          transform: scaleX(-1);
-        }
-      }
-    }
-  }
-
-  .tableContainer {
-    th > div {
-      text-align: right;
-    }
-    td {
-      text-align: right;
-      .serviceColor {
-        margin-right: 0;
-        margin-left: layout.$spacing-03;
-      }
-      button {
-        padding: layout.$spacing-03 0 layout.$spacing-03 layout.$spacing-05;
-        text-align: right;
-      }
-    }
-  }
-
-  .seeAllLink {
-    left: -22rem;
-  }
-}
