--- packages/esm-appointments-app/src/home/check-in-modal/check-in-modal.scss
@@ -1,12 +1,0 @@
-@use '@carbon/styles/scss/spacing';
-@use '@carbon/styles/scss/type';
-
-.checkInTime {
-  display: flex;
-  align-items: center;
-}
-
-.checkInLabel {
-  margin-right: spacing.$spacing-03;
-  font-weight: bold;
-}
