--- packages/esm-appointments-app/src/home/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import AppointmentsBaseTable from './appointments-base-table.component';
-
-const HomeAppointments = () => {
-  return <AppointmentsBaseTable />;
-};
-
-export default HomeAppointments;
