--- packages/esm-appointments-app/src/home/mock-appointments-data.ts
@@ -1,372 +1,0 @@
-export const appointmentsData = [
-  {
-    uuid: 'aed5f250-3988-4372-bb01-90b87e6c3d96',
-    appointmentNumber: '0000',
-    patient: {
-      identifier: 'SEM203001',
-      name: 'MARIAM MICHAEL MSAFIRI',
-      uuid: 'bf379289-62b7-47b3-ab7b-5186cb6fc46c',
-    },
-    service: {
-      appointmentServiceId: 1,
-      name: 'Cardiology',
-      description: null,
-      speciality: {
-        name: 'Cardiology',
-        uuid: '10a45eb9-e0b1-11e7-aec6-02e6b64603ba',
-      },
-      startTime: '',
-      endTime: '',
-      maxAppointmentsLimit: null,
-      durationMins: null,
-      location: {},
-      uuid: '7b200662-e21a-488c-8274-4eff81117608',
-      color: '#006400',
-      initialAppointmentStatus: null,
-      creatorName: null,
-    },
-    serviceType: null,
-    provider: null,
-    location: {
-      name: 'OPD-1',
-      uuid: 'c58e12ed-3f12-11e4-adec-0800271c1b75',
-    },
-    startDateTime: 1662671100000,
-    endDateTime: 1662672900000,
-    appointmentKind: 'WalkIn',
-    status: 'CheckedIn',
-    comments: 'Cardiology appointment',
-    additionalInfo: null,
-    teleconsultation: null,
-    providers: [],
-    voided: false,
-    extensions: {
-      patientEmailDefined: false,
-    },
-    teleconsultationLink: null,
-    recurring: false,
-  },
-  {
-    uuid: '4a94140a-edd1-4778-802e-83f330f448a4',
-    appointmentNumber: '0000',
-    patient: {
-      identifier: 'SEM203001',
-      name: 'MARIAM MICHAEL MSAFIRI',
-      uuid: 'bf379289-62b7-47b3-ab7b-5186cb6fc46c',
-    },
-    service: {
-      appointmentServiceId: 1,
-      name: 'Cardiology',
-      description: null,
-      speciality: {
-        name: 'Cardiology',
-        uuid: '10a45eb9-e0b1-11e7-aec6-02e6b64603ba',
-      },
-      startTime: '',
-      endTime: '',
-      maxAppointmentsLimit: null,
-      durationMins: null,
-      location: {},
-      uuid: '7b200662-e21a-488c-8274-4eff81117608',
-      color: '#DC143C',
-      initialAppointmentStatus: null,
-      creatorName: null,
-    },
-    serviceType: null,
-    provider: null,
-    location: {
-      name: 'OPD-1',
-      uuid: 'c58e12ed-3f12-11e4-adec-0800271c1b75',
-    },
-    startDateTime: 1662674700000,
-    endDateTime: 1662676500000,
-    appointmentKind: 'WalkIn',
-    status: 'Scheduled',
-    comments: 'Second appointment',
-    additionalInfo: null,
-    teleconsultation: null,
-    providers: [],
-    voided: false,
-    extensions: {
-      patientEmailDefined: false,
-    },
-    teleconsultationLink: null,
-    recurring: false,
-  },
-  {
-    uuid: 'e8561d5f-a157-4f71-85db-a0fa7aa05d6f',
-    appointmentNumber: '0000',
-    patient: {
-      identifier: 'SEM203001',
-      name: 'MARIAM MICHAEL MSAFIRI',
-      uuid: 'bf379289-62b7-47b3-ab7b-5186cb6fc46c',
-    },
-    service: {
-      appointmentServiceId: 1,
-      name: 'Cardiology',
-      description: null,
-      speciality: {
-        name: 'Cardiology',
-        uuid: '10a45eb9-e0b1-11e7-aec6-02e6b64603ba',
-      },
-      startTime: '',
-      endTime: '',
-      maxAppointmentsLimit: null,
-      durationMins: null,
-      location: {},
-      uuid: '7b200662-e21a-488c-8274-4eff81117608',
-      color: '#006400',
-      initialAppointmentStatus: null,
-      creatorName: null,
-    },
-    serviceType: null,
-    provider: null,
-    location: {
-      name: 'OPD-1',
-      uuid: 'c58e12ed-3f12-11e4-adec-0800271c1b75',
-    },
-    startDateTime: 1662678300000,
-    endDateTime: 1662680100000,
-    appointmentKind: 'WalkIn',
-    status: 'Scheduled',
-    comments: 'Third appointment',
-    additionalInfo: null,
-    teleconsultation: null,
-    providers: [],
-    voided: false,
-    extensions: {
-      patientEmailDefined: false,
-    },
-    teleconsultationLink: null,
-    recurring: false,
-  },
-  {
-    uuid: 'c1b14d42-49b3-4143-891c-03530bc812cc',
-    appointmentNumber: '0000',
-    patient: {
-      identifier: 'SEM203001',
-      name: 'MARIAM MICHAEL MSAFIRI',
-      uuid: 'bf379289-62b7-47b3-ab7b-5186cb6fc46c',
-    },
-    service: {
-      appointmentServiceId: 1,
-      name: 'Cardiology',
-      description: null,
-      speciality: {
-        name: 'Cardiology',
-        uuid: '10a45eb9-e0b1-11e7-aec6-02e6b64603ba',
-      },
-      startTime: '',
-      endTime: '',
-      maxAppointmentsLimit: null,
-      durationMins: null,
-      location: {},
-      uuid: '7b200662-e21a-488c-8274-4eff81117608',
-      color: '#00008B',
-      initialAppointmentStatus: null,
-      creatorName: null,
-    },
-    serviceType: null,
-    provider: null,
-    location: {
-      name: 'OPD-1',
-      uuid: 'c58e12ed-3f12-11e4-adec-0800271c1b75',
-    },
-    startDateTime: 1662685500000,
-    endDateTime: 1662687300000,
-    appointmentKind: 'WalkIn',
-    status: 'Completed',
-    comments: null,
-    additionalInfo: null,
-    teleconsultation: null,
-    providers: [],
-    voided: false,
-    extensions: {
-      patientEmailDefined: false,
-    },
-    teleconsultationLink: null,
-    recurring: false,
-  },
-  {
-    uuid: '6460ee51-542d-4afb-9cfe-6d3428446260',
-    appointmentNumber: '0000',
-    patient: {
-      identifier: 'GAN203022',
-      name: 'FSZoWMdgml FpJmfOMrcG tzDyuQLUkq',
-      uuid: 'a06a7722-acf9-4457-bacc-ca24774a9b69',
-    },
-    service: {
-      appointmentServiceId: 3,
-      name: 'Physiotherapy',
-      description: 'For test automation',
-      speciality: {},
-      startTime: '',
-      endTime: '',
-      maxAppointmentsLimit: null,
-      durationMins: null,
-      location: {},
-      uuid: '3a720936-90fb-4475-89bd-f155bf9752ba',
-      color: '#006400',
-      initialAppointmentStatus: null,
-      creatorName: null,
-    },
-    serviceType: null,
-    provider: null,
-    location: null,
-    startDateTime: 1662964200000,
-    endDateTime: 1662966000000,
-    appointmentKind: 'Scheduled',
-    status: 'Cancelled',
-    comments: null,
-    additionalInfo: null,
-    teleconsultation: null,
-    providers: [],
-    voided: false,
-    extensions: {
-      patientEmailDefined: false,
-    },
-    teleconsultationLink: null,
-    recurring: true,
-  },
-  {
-    uuid: 'ee49f558-c2d3-4965-8d28-614089623d59',
-    appointmentNumber: '0000',
-    patient: {
-      identifier: 'GAN203024',
-      name: 'Nirpesh kumar shamra',
-      uuid: 'd5dca632-1242-431c-842b-5e6bfa7f11b6',
-    },
-    service: {
-      appointmentServiceId: 1,
-      name: 'Cardiology',
-      description: null,
-      speciality: {
-        name: 'Cardiology',
-        uuid: '10a45eb9-e0b1-11e7-aec6-02e6b64603ba',
-      },
-      startTime: '',
-      endTime: '',
-      maxAppointmentsLimit: null,
-      durationMins: null,
-      location: {},
-      uuid: '7b200662-e21a-488c-8274-4eff81117608',
-      color: '#006400',
-      initialAppointmentStatus: null,
-      creatorName: null,
-    },
-    serviceType: null,
-    provider: null,
-    location: {
-      name: 'An Appointment Location',
-      uuid: '569f7e10-331d-4f7d-8b49-ffbd98b8f928',
-    },
-    startDateTime: 1663001520000,
-    endDateTime: 1663003320000,
-    appointmentKind: 'Scheduled',
-    status: 'Completed',
-    comments: null,
-    additionalInfo: null,
-    teleconsultation: null,
-    providers: [],
-    voided: false,
-    extensions: {
-      patientEmailDefined: false,
-    },
-    teleconsultationLink: null,
-    recurring: false,
-  },
-  {
-    uuid: '05b44534-ed12-4efe-9401-b5b5a9e0b05b',
-    appointmentNumber: '0000',
-    patient: {
-      identifier: 'GAN203025',
-      name: 'Krishna kumar sharma',
-      uuid: '7da5cd02-130e-4b8e-ad1b-d47797103798',
-    },
-    service: {
-      appointmentServiceId: 1,
-      name: 'Cardiology',
-      description: null,
-      speciality: {
-        name: 'Cardiology',
-        uuid: '10a45eb9-e0b1-11e7-aec6-02e6b64603ba',
-      },
-      startTime: '',
-      endTime: '',
-      maxAppointmentsLimit: null,
-      durationMins: null,
-      location: {},
-      uuid: '7b200662-e21a-488c-8274-4eff81117608',
-      color: '#006400',
-      initialAppointmentStatus: null,
-      creatorName: null,
-    },
-    serviceType: null,
-    provider: null,
-    location: {
-      name: 'Unknown Location',
-      uuid: '8d6c993e-c2cc-11de-8d13-0010c6dffd0f',
-    },
-    startDateTime: 1662997920000,
-    endDateTime: 1662999720000,
-    appointmentKind: 'WalkIn',
-    status: 'Cancelled',
-    comments: null,
-    additionalInfo: null,
-    teleconsultation: null,
-    providers: [],
-    voided: false,
-    extensions: {
-      patientEmailDefined: false,
-    },
-    teleconsultationLink: null,
-    recurring: false,
-  },
-  {
-    uuid: '05b44534-ed12-4efe-9401-b5b5a9e0b05c',
-    appointmentNumber: '0000',
-    patient: {
-      identifier: 'GAN203025',
-      name: 'kumar sharma',
-      uuid: '7da5cd02-130e-4b8e-ad1b-d47797103798',
-    },
-    service: {
-      appointmentServiceId: 1,
-      name: 'Cardiology',
-      description: null,
-      speciality: {
-        name: 'Cardiology',
-        uuid: '10a45eb9-e0b1-11e7-aec6-02e6b64603ba',
-      },
-      startTime: '',
-      endTime: '',
-      maxAppointmentsLimit: null,
-      durationMins: null,
-      location: {},
-      uuid: '7b200662-e21a-488c-8274-4eff81117608',
-      color: '#006400',
-      initialAppointmentStatus: null,
-      creatorName: null,
-    },
-    serviceType: null,
-    provider: null,
-    location: {
-      name: 'Unknown Location',
-      uuid: '8d6c993e-c2cc-11de-8d13-0010c6dffd0f',
-    },
-    startDateTime: 1662997920000,
-    endDateTime: 1662999720000,
-    appointmentKind: 'WalkIn',
-    status: 'Missed',
-    comments: null,
-    additionalInfo: null,
-    teleconsultation: null,
-    providers: [],
-    voided: false,
-    extensions: {
-      patientEmailDefined: false,
-    },
-    teleconsultationLink: null,
-    recurring: false,
-  },
-];
