+++ packages/esm-appointments-app/translations/en.json
@@ -3,28 +3,21 @@
   "actions": "Actions",
   "add": "Add",
   "addAppointment": "Add Appointment",
-  "addNewAppointment": "Add new appointment",
   "addPatientToQueue": "Add patient to queue",
   "age": "Age",
   "all": "All",
   "allDay": "All day",
-  "appointment": "appointment",
   "appointmentCancelError": "Error cancelling appointment",
   "appointmentCancelled": "Appointment Cancelled",
   "appointmentCancelledSuccessfully": "Appointment cancelled successfully",
-  "appointmentCheckedIn": "Appointment Checked-in",
-  "appointmentCheckInError": "Error checking in appointment",
   "appointmentColor": "Appointment color",
   "appointmentEdited": "Appointment edited",
   "appointmentEditError": "Error editing appointment",
   "appointmentForm": "Appointments Form",
   "appointmentFormError": "Error scheduling appointment",
   "appointmentHistory": "Appointment History",
-  "appointmentMarkedAsMissed": "It has been successfully marked as Missed",
   "appointmentMetrics": "Appointment metrics",
   "appointmentMetricsLoadError": "",
-  "appointmentMissed": "Appointment Missed",
-  "appointmentMissedError": "Error marking appointment as Missed",
   "appointmentNoteLabel": "Write an additional note",
   "appointmentNotePlaceholder": "Write any additional points here",
   "appointmentNotes": "Appointment Notes",
@@ -36,7 +29,6 @@
   "appointmentServiceCreate": "Appointment service created successfully",
   "appointmentServiceName": "Appointment service name",
   "appointmentStatus": "Appointment status",
-  "appointmentSuccessfullyCheckedIn": "It has been checked-in successfully",
   "appointmentToFulfill": "Select appointment(s) to fulfill",
   "appointmentType": "Appointment type",
   "appointmentType_title": "Appointment Type",
@@ -47,26 +39,19 @@
   "cancelAppointment": "Cancel Appointment",
   "cancelAppointmentModalConfirmationText": "Are you sure you want to cancel this appointment?",
   "cancelled": "Cancelled",
-  "cancelledSuccessfully": "It has been cancelled successfully",
   "checkedIn": "Checked in",
   "checkedOut": "Checked out",
-  "checkFilters": "Check the filters above",
   "checkIn": "Check In",
-  "checkInAppointment": "Are you sure you want to mark appointment as Checked-in?",
-  "checkinTime": "Check-in time",
   "checkOut": "Check out",
   "chooseAppointmentType": "Choose appointment type",
   "chooseLocation": "Choose a location",
   "chooseProvider": "Choose a provider",
   "chooseService": "Select service",
-  "complete": "Complete",
   "completed": "Completed",
-  "createAppointment": "Create appointment",
   "createAppointmentService": "Create appointment services",
   "createNewAppointment": "Create new appointment",
   "date": "Date",
   "date&Time": "Date & time",
-  "dateAndTime": "Date and time",
   "dateAndTimeOfVisit": "Date and time of visit",
   "dateOfBirth": "Date of birth",
   "dateTime": "Date & Time",
@@ -84,8 +69,6 @@
   "encounterType": "Encounter type",
   "endDate": "End date",
   "endTime": "End Time",
-  "errorCancellingAppointment": "Error cancelling the appointment",
-  "errorCheckingAppoitment": "Error checking in the appointment",
   "errorCreatingAppointmentService": "Error creating appointment service",
   "expected": "Expected",
   "gender": "Gender",
@@ -100,10 +83,7 @@
   "medications": "Medications",
   "missed": "Missed",
   "missingVisitType": "Missing visit type",
-  "name": "Name",
   "no": "No",
-  "noAppointmentsScheduledForTodayToDisplay": "There are no appointments scheduled for today to display for this location",
-  "noAppointmentsToDisplay": "There are no appointments to display for this location",
   "noContent": "No Content",
   "noCurrentAppointments": "There are no appointments scheduled for today to display for this patient",
   "noEncountersFound": "No encounters found",
@@ -115,7 +95,6 @@
   "notes": "Notes",
   "noUpcomingAppointments": "No upcoming appointments found",
   "noUpcomingAppointmentsForPatient": "There are no upcoming appointments to display for this patient",
-  "or": "or",
   "past": "Past",
   "patientDetails": "Patient Details",
   "patientName": "Patient name",
@@ -127,17 +106,14 @@
   "providersBooked": "Providers booked: {{time}}",
   "queueAddedSuccessfully": "",
   "queueEntryError": "Error adding patient to the queue",
-  "reasonForChanges": "Reason For Changes",
   "recurringAppointment": "Recurring Appointment",
   "repeatEvery": "Repeat every",
-  "revertingAppointmentStatus": "Reverting appointment status",
   "save": "Save",
   "saveAndClose": "Save and close",
   "scheduled": "Scheduled",
   "scheduledAppointments": "Scheduled appointments",
   "search": "Search",
   "searchForAVisitType": "Search for a visit type",
-  "seeAllAppointments": "See all appointments",
   "selectAppointmentStatus": "Select status",
   "selectAppointmentType": "Select an appointment type",
   "selectFacility": "Select a facility",
@@ -155,16 +131,14 @@
   "startVisitError": "Error starting visit",
   "time": "Time",
   "today": "Today",
-  "todaysAppointments": "Today's Appointments",
+  "todays": "Today's",
   "total": "Total",
-  "undo": "Undo",
   "unscheduled": "Unscheduled",
   "unscheduledAppointments": "Unscheduled appointments",
   "unscheduledAppointments_lower": "unscheduled appointments",
   "upcoming": "Upcoming",
   "upcomingAppointments": "Upcoming appointments",
   "view": "View",
-  "viewCalendar": "View Calendar",
   "visitLocation": "Visit Location",
   "visitStarted": "Visit started",
   "visitType": "Visit Type",
